refactor(gba/apu): move each channel struct into its own file
- Add fifo.rs: FifoChannel + FIFO_CAPACITY, with 4 unit tests
- Add channel1.rs: Channel1 (pulse with sweep), uses DUTY_TABLE from apu.rs
- Add channel2.rs: Channel2 (pulse without sweep), uses DUTY_TABLE from apu.rs
- Add channel3.rs: Channel3 (wave playback) + SAMPLES_PER_BANK, with 2 unit tests
- Add channel4.rs: Channel4 (noise/LFSR) + DIVISORS, with 6 unit tests
- Update mod.rs: declare the five new submodules
- Update apu.rs: remove moved structs/constants, add use imports,
  mark DUTY_TABLE as pub(super), keep Apu orchestrator and FS_TABLE

Closes #2477
