+++ src/gba/apu/apu.rs
@@ -1,5 +1,11 @@
 //! GBA APU — dual-source audio: 4 DMG-legacy channels + 2 PCM FIFO channels.
 
+use super::channel1::Channel1;
+use super::channel2::Channel2;
+use super::channel3::Channel3;
+use super::channel4::Channel4;
+use super::fifo::FifoChannel;
+
 // ── Timing constants ──────────────────────────────────────────────────────────
 
 /// GBA CPU clock rate in Hz.
@@ -11,734 +17,12 @@ const FS_PERIOD: u32 = 32_768;
 // ── Duty cycle waveforms ──────────────────────────────────────────────────────
 
 /// 8-step duty-cycle waveforms (identical to DMG APU).
-const DUTY_TABLE: [[u8; 8]; 4] = [
+pub(super) const DUTY_TABLE: [[u8; 8]; 4] = [
     [0, 0, 0, 0, 0, 0, 0, 1], // 12.5 %
     [1, 0, 0, 0, 0, 0, 0, 1], // 25 %
     [1, 0, 0, 0, 0, 1, 1, 1], // 50 %
     [0, 1, 1, 1, 1, 1, 1, 0], // 75 %
 ];
-
-/// CH4 noise-clock divisors (in GB T-cycles; ×4 gives GBA cycles).
-const DIVISORS: [u32; 8] = [8, 16, 32, 48, 64, 80, 96, 112];
-
-// ── PCM FIFO ─────────────────────────────────────────────────────────────────
-
-/// GBA FIFO capacity: 32 bytes per GBATek spec.
-const FIFO_CAPACITY: usize = 32;
-
-/// PCM FIFO channel (A or B).
-#[derive(Debug, Clone, Default)]
-pub struct FifoChannel {
-    samples: std::collections::VecDeque<i8>,
-    /// The sample currently being output.
-    pub current: i8,
-}
-
-impl FifoChannel {
-    /// Push one 8-bit signed sample into the FIFO (ignored when full).
-    pub fn push(&mut self, sample: i8) {
-        if self.samples.len() < FIFO_CAPACITY {
-            self.samples.push_back(sample);
-        }
-    }
-
-    /// Push four bytes (one word) into the FIFO, low byte first.
-    pub fn push_word(&mut self, word: u32) {
-        for i in 0..4 {
-            self.push(((word >> (i * 8)) as u8) as i8);
-        }
-    }
-
-    /// Consume the next sample from the FIFO, updating `current`.
-    /// When the FIFO is empty the current sample holds.
-    pub fn advance(&mut self) {
-        if let Some(s) = self.samples.pop_front() {
-            self.current = s;
-        }
-    }
-
-    /// Clear FIFO and reset current sample to 0.
-    pub fn clear(&mut self) {
-        self.samples.clear();
-        self.current = 0;
-    }
-
-    /// Number of samples waiting in the FIFO.
-    pub fn len(&self) -> usize {
-        self.samples.len()
-    }
-
-    /// True when the FIFO is empty.
-    pub fn is_empty(&self) -> bool {
-        self.samples.is_empty()
-    }
-
-    /// Normalised output in `[-1.0, 1.0]`.
-    pub fn output(&self) -> f32 {
-        self.current as f32 / 128.0
-    }
-}
-
-// ── CH1 — Pulse + Sweep ───────────────────────────────────────────────────────
-
-/// Channel 1: pulse wave with frequency sweep.
-#[derive(Debug, Clone, Default)]
-pub struct Channel1 {
-    // Sweep (SOUND1CNT_L)
-    pub sweep_period: u8,
-    pub sweep_negate: bool,
-    pub sweep_shift: u8,
-
-    // Tone (SOUND1CNT_H)
-    pub duty: u8,
-    pub length_counter: u16,
-    pub init_volume: u8,
-    pub env_add: bool,
-    pub env_period: u8,
-
-    // Frequency + control (SOUND1CNT_X)
-    pub freq: u16,
-    pub length_en: bool,
-
-    // Internal state
-    pub active: bool,
-    pub dac_on: bool,
-    pub duty_pos: u8,
-    pub freq_timer: u32,
-    pub volume: u8,
-    pub env_timer: u8,
-    pub sweep_timer: u8,
-    /// Shadow copy of frequency used by sweep.
-    pub sweep_shadow: u16,
-    pub sweep_enabled: bool,
-}
-
-impl Channel1 {
-    /// Analogue output in `[0.0, 1.0]`; 0 when inactive.
-    pub fn output(&self) -> f32 {
-        if !self.active || !self.dac_on {
-            return 0.0;
-        }
-        let bit = DUTY_TABLE[self.duty as usize][self.duty_pos as usize];
-        if bit == 1 {
-            self.volume as f32 / 15.0
-        } else {
-            0.0
-        }
-    }
-
-    /// Advance channel by `cycles` GBA cycles (frequency timer only).
-    pub fn tick(&mut self, cycles: u32) {
-        if !self.active {
-            return;
-        }
-        let period = (2048_u32.wrapping_sub(self.freq as u32)) * 16;
-        if period == 0 {
-            return;
-        }
-        let mut rem = cycles;
-        while rem > 0 {
-            if self.freq_timer == 0 {
-                self.freq_timer = period;
-            }
-            let advance = rem.min(self.freq_timer);
-            self.freq_timer -= advance;
-            rem -= advance;
-            if self.freq_timer == 0 {
-                self.duty_pos = (self.duty_pos + 1) & 7;
-                self.freq_timer = period;
-            }
-        }
-    }
-
-    /// Clock the length counter (FS steps 0, 2, 4, 6).
-    pub fn clock_length(&mut self) {
-        if !self.length_en || self.length_counter == 0 {
-            return;
-        }
-        self.length_counter -= 1;
-        if self.length_counter == 0 {
-            self.active = false;
-        }
-    }
-
-    /// Clock the volume envelope (FS step 7).
-    pub fn clock_envelope(&mut self) {
-        if self.env_period == 0 {
-            return;
-        }
-        if self.env_timer > 0 {
-            self.env_timer -= 1;
-        }
-        if self.env_timer == 0 {
-            self.env_timer = self.env_period;
-            if self.env_add && self.volume < 15 {
-                self.volume += 1;
-            } else if !self.env_add && self.volume > 0 {
-                self.volume -= 1;
-            }
-        }
-    }
-
-    /// Clock the frequency sweep (FS steps 2 and 6).
-    pub fn clock_sweep(&mut self) {
-        if self.sweep_timer > 0 {
-            self.sweep_timer -= 1;
-        }
-        if self.sweep_timer == 0 {
-            self.sweep_timer = if self.sweep_period > 0 {
-                self.sweep_period
-            } else {
-                8
-            };
-            if self.sweep_enabled && self.sweep_period > 0 {
-                let new_freq = self.calc_sweep();
-                if new_freq > 2047 {
-                    self.active = false;
-                } else if self.sweep_shift > 0 {
-                    self.sweep_shadow = new_freq;
-                    self.freq = new_freq;
-                    // Overflow check with new shadow.
-                    if self.calc_sweep() > 2047 {
-                        self.active = false;
-                    }
-                }
-            }
-        }
-    }
-
-    fn calc_sweep(&self) -> u16 {
-        if self.sweep_negate {
-            self.sweep_shadow
-                .wrapping_sub(self.sweep_shadow >> self.sweep_shift)
-        } else {
-            self.sweep_shadow
-                .wrapping_add(self.sweep_shadow >> self.sweep_shift)
-        }
-    }
-
-    /// Trigger (write 1 to reset/trigger bit).
-    pub fn trigger(&mut self) {
-        self.active = self.dac_on;
-        if self.length_counter == 0 {
-            self.length_counter = 64;
-        }
-        let period = (2048_u32.wrapping_sub(self.freq as u32)) * 16;
-        self.freq_timer = period;
-        self.volume = self.init_volume;
-        self.env_timer = self.env_period;
-        // Sweep initialisation.
-        self.sweep_shadow = self.freq;
-        self.sweep_timer = if self.sweep_period > 0 {
-            self.sweep_period
-        } else {
-            8
-        };
-        self.sweep_enabled = self.sweep_period > 0 || self.sweep_shift > 0;
-        // Overflow check on trigger.
-        if self.sweep_enabled && self.sweep_shift > 0 && self.calc_sweep() > 2047 {
-            self.active = false;
-        }
-    }
-
-    // ── Register writes ───────────────────────────────────────────────────
-
-    /// SOUND1CNT_L: sweep register (bits 6-4 period, bit 3 negate, bits 2-0 shift).
-    pub fn write_cnt_l(&mut self, val: u16) {
-        self.sweep_shift = (val & 0x07) as u8;
-        self.sweep_negate = (val & 0x08) != 0;
-        self.sweep_period = ((val >> 4) & 0x07) as u8;
-    }
-
-    /// SOUND1CNT_H: tone / envelope (bits 5-0 length, 7-6 duty, 10-8 env period,
-    /// 11 env add, 15-12 init volume).
-    pub fn write_cnt_h(&mut self, val: u16) {
-        self.length_counter = 64 - (val & 0x3F);
-        self.duty = ((val >> 6) & 0x03) as u8;
-        self.env_period = ((val >> 8) & 0x07) as u8;
-        self.env_add = (val & 0x0800) != 0;
-        self.init_volume = ((val >> 12) & 0x0F) as u8;
-        self.dac_on = (val & 0xF800) != 0;
-        if !self.dac_on {
-            self.active = false;
-        }
-    }
-
-    /// SOUND1CNT_X: frequency and control (bits 10-0 freq, 14 length-enable,
-    /// 15 trigger/reset — write-only).
-    pub fn write_cnt_x(&mut self, val: u16) {
-        self.freq = val & 0x7FF;
-        self.length_en = (val & 0x4000) != 0;
-        if val & 0x8000 != 0 {
-            self.trigger();
-        }
-    }
-
-    /// Power-off reset.
-    pub fn power_off(&mut self) {
-        *self = Self::default();
-    }
-}
-
-// ── CH2 — Pulse (no sweep) ────────────────────────────────────────────────────
-
-/// Channel 2: pulse wave without sweep.
-#[derive(Debug, Clone, Default)]
-pub struct Channel2 {
-    pub duty: u8,
-    pub length_counter: u16,
-    pub init_volume: u8,
-    pub env_add: bool,
-    pub env_period: u8,
-    pub freq: u16,
-    pub length_en: bool,
-
-    pub active: bool,
-    pub dac_on: bool,
-    pub duty_pos: u8,
-    pub freq_timer: u32,
-    pub volume: u8,
-    pub env_timer: u8,
-}
-
-impl Channel2 {
-    pub fn output(&self) -> f32 {
-        if !self.active || !self.dac_on {
-            return 0.0;
-        }
-        let bit = DUTY_TABLE[self.duty as usize][self.duty_pos as usize];
-        if bit == 1 {
-            self.volume as f32 / 15.0
-        } else {
-            0.0
-        }
-    }
-
-    pub fn tick(&mut self, cycles: u32) {
-        if !self.active {
-            return;
-        }
-        let period = (2048_u32.wrapping_sub(self.freq as u32)) * 16;
-        if period == 0 {
-            return;
-        }
-        let mut rem = cycles;
-        while rem > 0 {
-            if self.freq_timer == 0 {
-                self.freq_timer = period;
-            }
-            let advance = rem.min(self.freq_timer);
-            self.freq_timer -= advance;
-            rem -= advance;
-            if self.freq_timer == 0 {
-                self.duty_pos = (self.duty_pos + 1) & 7;
-                self.freq_timer = period;
-            }
-        }
-    }
-
-    pub fn clock_length(&mut self) {
-        if !self.length_en || self.length_counter == 0 {
-            return;
-        }
-        self.length_counter -= 1;
-        if self.length_counter == 0 {
-            self.active = false;
-        }
-    }
-
-    pub fn clock_envelope(&mut self) {
-        if self.env_period == 0 {
-            return;
-        }
-        if self.env_timer > 0 {
-            self.env_timer -= 1;
-        }
-        if self.env_timer == 0 {
-            self.env_timer = self.env_period;
-            if self.env_add && self.volume < 15 {
-                self.volume += 1;
-            } else if !self.env_add && self.volume > 0 {
-                self.volume -= 1;
-            }
-        }
-    }
-
-    fn trigger(&mut self) {
-        self.active = self.dac_on;
-        if self.length_counter == 0 {
-            self.length_counter = 64;
-        }
-        let period = (2048_u32.wrapping_sub(self.freq as u32)) * 16;
-        self.freq_timer = period;
-        self.volume = self.init_volume;
-        self.env_timer = self.env_period;
-    }
-
-    // ── Register writes ───────────────────────────────────────────────────
-
-    /// SOUND2CNT_L: duty/length/envelope.
-    pub fn write_cnt_l(&mut self, val: u16) {
-        self.length_counter = 64 - (val & 0x3F);
-        self.duty = ((val >> 6) & 0x03) as u8;
-        self.env_period = ((val >> 8) & 0x07) as u8;
-        self.env_add = (val & 0x0800) != 0;
-        self.init_volume = ((val >> 12) & 0x0F) as u8;
-        self.dac_on = (val & 0xF800) != 0;
-        if !self.dac_on {
-            self.active = false;
-        }
-    }
-
-    /// SOUND2CNT_H: frequency and trigger.
-    pub fn write_cnt_h(&mut self, val: u16) {
-        self.freq = val & 0x7FF;
-        self.length_en = (val & 0x4000) != 0;
-        if val & 0x8000 != 0 {
-            self.trigger();
-        }
-    }
-
-    pub fn power_off(&mut self) {
-        *self = Self::default();
-    }
-}
-
-// ── CH3 — Wave ────────────────────────────────────────────────────────────────
-
-/// Number of 4-bit samples per wave RAM bank.
-const SAMPLES_PER_BANK: u8 = 32;
-
-/// Channel 3: wave playback (32 × 4-bit samples, or 64 × 4-bit samples in two-bank mode).
-#[derive(Debug, Clone, Default)]
-pub struct Channel3 {
-    pub dac_on: bool,
-    /// Bit 5 of SOUND3CNT_L: 0 = one bank (32 samples), 1 = two banks (64 samples).
-    pub two_banks: bool,
-    /// Bit 6 of SOUND3CNT_L: selects which bank (0 or 1) is played back.
-    /// The OTHER bank is accessible via Wave RAM register reads/writes.
-    pub bank_select: bool,
-    pub length_counter: u16,
-    pub output_level: u8, // 0=mute, 1=100%, 2=50%, 3=25%
-    /// Bit 15 of SOUND3CNT_H: when true, forces 75% volume regardless of output_level.
-    pub force_volume: bool,
-    pub freq: u16,
-    pub length_en: bool,
-
-    pub active: bool,
-    pub wave_pos: u8, // 0-31 (single bank) or 0-63 (two banks)
-    pub freq_timer: u32,
-    /// Wave RAM: 2 banks × 16 bytes = 32 bytes total (64 × 4-bit samples).
-    pub wave_ram: [[u8; 16]; 2],
-    /// Current 4-bit output nibble.
-    pub current_sample: u8,
-}
-
-impl Channel3 {
-    pub fn output(&self) -> f32 {
-        if !self.active || !self.dac_on {
-            return 0.0;
-        }
-        // Bit 15 of SOUND3CNT_H: force 75% volume regardless of output_level.
-        if self.force_volume {
-            return ((self.current_sample as u16 * 3) / 4) as f32 / 15.0;
-        }
-        let shift: u8 = match self.output_level {
-            1 => 0, // 100 %
-            2 => 1, // 50 %
-            3 => 2, // 25 %
-            _ => return 0.0,
-        };
-        (self.current_sample >> shift) as f32 / 15.0
-    }
-
-    /// Advance channel by `cycles` GBA cycles.
-    /// Each wave_pos step takes (2048 − freq) × 8 GBA cycles.
-    pub fn tick(&mut self, cycles: u32) {
-        if !self.active {
-            return;
-        }
-        let period = (2048_u32.wrapping_sub(self.freq as u32)) * 8;
-        if period == 0 {
-            return;
-        }
-        // Use a bitmask to wrap wave_pos: single-bank → mask 0x1F (& 31), two-bank → 0x3F (& 63).
-        let pos_mask: u8 = if self.two_banks { 0x3F } else { 0x1F };
-        let mut rem = cycles;
-        while rem > 0 {
-            if self.freq_timer == 0 {
-                self.freq_timer = period;
-            }
-            let advance = rem.min(self.freq_timer);
-            self.freq_timer -= advance;
-            rem -= advance;
-            if self.freq_timer == 0 {
-                self.wave_pos = (self.wave_pos + 1) & pos_mask;
-                // Samples 0..(SAMPLES_PER_BANK-1) come from bank_select; the rest from the other bank.
-                let bank = if self.wave_pos < SAMPLES_PER_BANK {
-                    self.bank_select as usize
-                } else {
-                    (self.bank_select as usize) ^ 1
-                };
-                let pos_in_bank = self.wave_pos & (SAMPLES_PER_BANK - 1);
-                let byte = self.wave_ram[bank][(pos_in_bank / 2) as usize];
-                self.current_sample = if pos_in_bank & 1 == 0 {
-                    (byte >> 4) & 0x0F // high nibble
-                } else {
-                    byte & 0x0F // low nibble
-                };
-                self.freq_timer = period;
-            }
-        }
-    }
-
-    pub fn clock_length(&mut self) {
-        if !self.length_en || self.length_counter == 0 {
-            return;
-        }
-        self.length_counter -= 1;
-        if self.length_counter == 0 {
-            self.active = false;
-        }
-    }
-
-    fn trigger(&mut self) {
-        self.active = self.dac_on;
-        if self.length_counter == 0 {
-            self.length_counter = 256;
-        }
-        let period = (2048_u32.wrapping_sub(self.freq as u32)) * 8;
-        self.freq_timer = period;
-        self.wave_pos = 0;
-        // Read first sample from the playing bank (position 0 = high nibble of byte 0).
-        let bank = self.bank_select as usize;
-        self.current_sample = (self.wave_ram[bank][0] >> 4) & 0x0F;
-    }
-
-    // ── Register writes ───────────────────────────────────────────────────
-
-    /// SOUND3CNT_L: wave RAM dimension (bit 5), bank select (bit 6), DAC enable (bit 7).
-    pub fn write_cnt_l(&mut self, val: u16) {
-        self.two_banks = (val & 0x0020) != 0;
-        self.bank_select = (val & 0x0040) != 0;
-        self.dac_on = (val & 0x0080) != 0;
-        if !self.dac_on {
-            self.active = false;
-        }
-    }
-
-    /// SOUND3CNT_H: sound length (bits 7-0), output level (bits 14-13), force volume (bit 15).
-    pub fn write_cnt_h(&mut self, val: u16) {
-        // Lower byte = sound length (0-255, counter = 256 - length)
-        self.length_counter = 256 - (val & 0x00FF);
-        // Bits 14-13: output level
-        self.output_level = ((val >> 13) & 0x03) as u8;
-        // Bit 15: force 75% volume regardless of output_level
-        self.force_volume = (val & 0x8000) != 0;
-    }
-
-    /// SOUND3CNT_X: frequency and trigger.
-    pub fn write_cnt_x(&mut self, val: u16) {
-        self.freq = val & 0x7FF;
-        self.length_en = (val & 0x4000) != 0;
-        if val & 0x8000 != 0 {
-            self.trigger();
-        }
-    }
-
-    /// Write one byte to wave RAM (address offset 0x00–0x0F from wave RAM base).
-    /// Always accesses the bank NOT currently selected for playback.
-    pub fn write_wave_ram(&mut self, offset: usize, val: u8) {
-        let other_bank = (self.bank_select as usize) ^ 1;
-        if offset < 16 {
-            self.wave_ram[other_bank][offset] = val;
-        }
-    }
-
-    /// Read one byte from wave RAM.
-    /// Always accesses the bank NOT currently selected for playback.
-    pub fn read_wave_ram(&self, offset: usize) -> u8 {
-        let other_bank = (self.bank_select as usize) ^ 1;
-        if offset < 16 {
-            self.wave_ram[other_bank][offset]
-        } else {
-            0xFF
-        }
-    }
-
-    pub fn power_off(&mut self) {
-        let wave_ram = self.wave_ram;
-        *self = Self::default();
-        self.wave_ram = wave_ram; // both wave RAM banks survive power-off
-    }
-}
-
-// ── CH4 — Noise ───────────────────────────────────────────────────────────────
-
-/// Channel 4: noise generator (15-bit or 7-bit LFSR).
-#[derive(Debug, Clone)]
-pub struct Channel4 {
-    pub init_volume: u8,
-    pub env_add: bool,
-    pub env_period: u8,
-    pub clock_shift: u8,
-    pub lfsr_7bit: bool,
-    pub divisor_code: u8,
-    pub length_counter: u16,
-    pub length_en: bool,
-
-    pub active: bool,
-    pub dac_on: bool,
-    pub lfsr: u16,
-    pub freq_timer: u32,
-    pub volume: u8,
-    pub env_timer: u8,
-}
-
-impl Default for Channel4 {
-    fn default() -> Self {
-        Self {
-            init_volume: 0,
-            env_add: false,
-            env_period: 0,
-            clock_shift: 0,
-            lfsr_7bit: false,
-            divisor_code: 0,
-            length_counter: 0,
-            length_en: false,
-            active: false,
-            dac_on: false,
-            // GBATek: 15-bit mode initial value is 0x4000.
-            lfsr: 0x4000,
-            freq_timer: 0,
-            volume: 0,
-            env_timer: 0,
-        }
-    }
-}
-
-impl Channel4 {
-    pub fn output(&self) -> f32 {
-        if !self.active || !self.dac_on {
-            return 0.0;
-        }
-        // LFSR bit 0 low → output high.
-        if self.lfsr & 0x01 == 0 {
-            self.volume as f32 / 15.0
-        } else {
-            0.0
-        }
-    }
-
-    fn freq_period(&self) -> u32 {
-        DIVISORS[self.divisor_code as usize] * (1_u32 << self.clock_shift) * 4
-    }
-
-    /// Advance channel by `cycles` GBA cycles.
-    pub fn tick(&mut self, cycles: u32) {
-        if !self.active {
-            return;
-        }
-        let period = self.freq_period();
-        if period == 0 {
-            return;
-        }
-        let mut rem = cycles;
-        while rem > 0 {
-            if self.freq_timer == 0 {
-                self.freq_timer = period;
-            }
-            let advance = rem.min(self.freq_timer);
-            self.freq_timer -= advance;
-            rem -= advance;
-            if self.freq_timer == 0 {
-                self.clock_lfsr();
-                self.freq_timer = period;
-            }
-        }
-    }
-
-    /// Clock the LFSR (exposed for testing).
-    ///
-    /// Implements the GBATek Galois LFSR form:
-    ///   15-bit: X = X SHR 1; IF carry THEN X = X XOR 6000h
-    ///   7-bit:  X = X SHR 1; IF carry THEN X = X XOR 0060h
-    pub fn clock_lfsr(&mut self) {
-        let carry = self.lfsr & 1;
-        self.lfsr >>= 1;
-        if carry != 0 {
-            self.lfsr ^= if self.lfsr_7bit { 0x0060 } else { 0x6000 };
-        }
-    }
-
-    pub fn clock_length(&mut self) {
-        if !self.length_en || self.length_counter == 0 {
-            return;
-        }
-        self.length_counter -= 1;
-        if self.length_counter == 0 {
-            self.active = false;
-        }
-    }
-
-    pub fn clock_envelope(&mut self) {
-        if self.env_period == 0 {
-            return;
-        }
-        if self.env_timer > 0 {
-            self.env_timer -= 1;
-        }
-        if self.env_timer == 0 {
-            self.env_timer = self.env_period;
-            if self.env_add && self.volume < 15 {
-                self.volume += 1;
-            } else if !self.env_add && self.volume > 0 {
-                self.volume -= 1;
-            }
-        }
-    }
-
-    fn trigger(&mut self) {
-        self.active = self.dac_on;
-        if self.length_counter == 0 {
-            self.length_counter = 64;
-        }
-        self.freq_timer = self.freq_period();
-        self.volume = self.init_volume;
-        self.env_timer = self.env_period;
-        // GBATek: 15-bit mode initial value 0x4000; 7-bit mode initial value 0x0040.
-        self.lfsr = if self.lfsr_7bit { 0x0040 } else { 0x4000 };
-    }
-
-    // ── Register writes ───────────────────────────────────────────────────
-
-    /// SOUND4CNT_L: length (5-0), envelope (10-8, 11, 15-12).
-    pub fn write_cnt_l(&mut self, val: u16) {
-        self.length_counter = 64 - (val & 0x3F);
-        self.env_period = ((val >> 8) & 0x07) as u8;
-        self.env_add = (val & 0x0800) != 0;
-        self.init_volume = ((val >> 12) & 0x0F) as u8;
-        self.dac_on = (val & 0xF800) != 0;
-        if !self.dac_on {
-            self.active = false;
-        }
-    }
-
-    /// SOUND4CNT_H: clock (7-4 shift, 3 LFSR mode, 2-0 divisor), 14 length-en, 15 trigger.
-    pub fn write_cnt_h(&mut self, val: u16) {
-        self.divisor_code = (val & 0x07) as u8;
-        self.lfsr_7bit = (val & 0x08) != 0;
-        self.clock_shift = ((val >> 4) & 0x0F) as u8;
-        self.length_en = (val & 0x4000) != 0;
-        if val & 0x8000 != 0 {
-            self.trigger();
-        }
-    }
-
-    pub fn power_off(&mut self) {
-        *self = Self::default();
-    }
-}
 
 // ── Frame Sequencer step table ────────────────────────────────────────────────
 
@@ -1443,53 +727,6 @@ mod tests {
         apu
     }
 
-    // ── FIFO channel tests ───────────────────────────────────────────────────
-
-    #[test]
-    fn test_fifo_push_and_advance() {
-        let mut fifo = FifoChannel::default();
-        fifo.push(42);
-        fifo.push(-10);
-        fifo.advance();
-        assert_eq!(fifo.current, 42);
-        fifo.advance();
-        assert_eq!(fifo.current, -10);
-    }
-
-    #[test]
-    fn test_fifo_capacity_limit() {
-        let mut fifo = FifoChannel::default();
-        for i in 0..40i8 {
-            fifo.push(i);
-        }
-        assert_eq!(
-            fifo.len(),
-            FIFO_CAPACITY,
-            "FIFO should cap at {FIFO_CAPACITY}"
-        );
-    }
-
-    #[test]
-    fn test_fifo_clear() {
-        let mut fifo = FifoChannel::default();
-        fifo.push(10);
-        fifo.clear();
-        assert!(fifo.is_empty());
-        assert_eq!(fifo.current, 0);
-    }
-
-    #[test]
-    fn test_fifo_output_normalised() {
-        let mut fifo = FifoChannel::default();
-        fifo.push(64);
-        fifo.advance();
-        let out = fifo.output();
-        assert!(
-            (out - 0.5).abs() < 1e-6,
-            "64 / 128 should be 0.5, got {out}"
-        );
-    }
-
     // ── PCM channel A test (issue test vector) ───────────────────────────────
 
     #[test]
@@ -1702,47 +939,6 @@ mod tests {
     }
 
     #[test]
-    fn test_ch3_force_volume_output_is_75_percent() {
-        // RED: when force_volume is set, output() must return (sample*3)/4 with truncation.
-        // Sample = 1 -> (1*3)/4 = 0, expected output = 0.
-        let ch3 = Channel3 {
-            dac_on: true,
-            active: true,
-            force_volume: true,
-            current_sample: 1,
-            output_level: 0, // would normally mute — force_volume overrides
-            ..Channel3::default()
-        };
-        let got = ch3.output();
-        let expected = ((1_u16 * 3) / 4) as f32 / 15.0;
-        assert!(
-            (got - expected).abs() < 1e-5,
-            "force_volume output mismatch: expected {expected}, got {got}"
-        );
-    }
-
-    #[test]
-    fn test_ch3_force_volume_overrides_output_level() {
-        // When force_volume is set and output_level=3 (25%), output should be 75%, not 25%.
-        // Sample = 12 → force 75%: 12 * 3/4 = 9; expected = 9/15 = 0.6
-        // Normal 25%: 12 >> 2 = 3; 3/15 = 0.2
-        let ch3 = Channel3 {
-            dac_on: true,
-            active: true,
-            force_volume: true,
-            current_sample: 12,
-            output_level: 3, // 25% — should be overridden
-            ..Channel3::default()
-        };
-        let got = ch3.output();
-        let expected = ((12_u16 * 3) / 4) as f32 / 15.0;
-        assert!(
-            (got - expected).abs() < 1e-5,
-            "force_volume should override output_level: expected {expected}, got {got}"
-        );
-    }
-
-    #[test]
     fn test_write8_sound3cnt_h_high_byte_does_not_clobber_length() {
         let mut apu = powered_apu();
         apu.write8(0x0400_0072, 0x20); // length = 32 -> counter = 224
@@ -1807,50 +1003,6 @@ mod tests {
     }
 
     #[test]
-    fn test_ch4_lfsr_galois_no_carry_step() {
-        // RED: Galois clock step when carry=0: only right shift, no XOR.
-        // Starting from 0x2000 (bit 13 set, bit 0 = 0 → carry = 0).
-        let mut ch4 = Channel4 {
-            lfsr: 0x2000,
-            ..Channel4::default()
-        };
-        ch4.clock_lfsr();
-        assert_eq!(ch4.lfsr, 0x1000, "Carry=0: should just shift right");
-    }
-
-    #[test]
-    fn test_ch4_lfsr_galois_carry_15bit() {
-        // RED: Galois clock step with carry=1 in 15-bit mode.
-        // State 0x0001: carry=1, shift → 0x0000, XOR 0x6000 → 0x6000.
-        let mut ch4 = Channel4 {
-            lfsr: 0x0001,
-            lfsr_7bit: false,
-            ..Channel4::default()
-        };
-        ch4.clock_lfsr();
-        assert_eq!(
-            ch4.lfsr, 0x6000,
-            "15-bit carry=1: 0x0001 >> 1 = 0x0000, XOR 0x6000 = 0x6000"
-        );
-    }
-
-    #[test]
-    fn test_ch4_lfsr_galois_carry_7bit() {
-        // RED: Galois clock step with carry=1 in 7-bit mode.
-        // State 0x0001: carry=1, shift → 0x0000, XOR 0x0060 → 0x0060.
-        let mut ch4 = Channel4 {
-            lfsr: 0x0001,
-            lfsr_7bit: true,
-            ..Channel4::default()
-        };
-        ch4.clock_lfsr();
-        assert_eq!(
-            ch4.lfsr, 0x0060,
-            "7-bit carry=1: 0x0001 >> 1 = 0x0000, XOR 0x0060 = 0x0060"
-        );
-    }
-
-    #[test]
     fn test_ch4_lfsr_sequence_15bit() {
         let mut apu = powered_apu();
 
@@ -1883,70 +1035,6 @@ mod tests {
                 apu.ch4.lfsr
             );
         }
-    }
-
-    #[test]
-    fn test_ch4_lfsr_sequence_7bit() {
-        let mut ch4 = Channel4 {
-            dac_on: true,
-            lfsr_7bit: true,
-            lfsr: 0x0040,
-            ..Channel4::default()
-        };
-
-        // Clock 7-bit LFSR manually using GBATek Galois form and verify.
-        let mut lfsr = 0x0040_u16;
-        fn clock_7bit(lfsr: &mut u16) {
-            let carry = *lfsr & 1;
-            *lfsr >>= 1;
-            if carry != 0 {
-                *lfsr ^= 0x0060;
-            }
-        }
-
-        for _ in 0..16 {
-            clock_7bit(&mut lfsr);
-            ch4.clock_lfsr();
-            assert_eq!(
-                ch4.lfsr, lfsr,
-                "7-bit LFSR mismatch: expected 0x{lfsr:04X}, got 0x{:04X}",
-                ch4.lfsr
-            );
-        }
-    }
-
-    #[test]
-    fn test_ch4_lfsr_15bit_maximal_length() {
-        // Verify the 15-bit Galois LFSR has maximum period of 32767 (= 2^15 - 1).
-        let mut lfsr = 0x4000_u16;
-        for _ in 0..32767 {
-            let carry = lfsr & 1;
-            lfsr >>= 1;
-            if carry != 0 {
-                lfsr ^= 0x6000;
-            }
-        }
-        assert_eq!(
-            lfsr, 0x4000,
-            "15-bit LFSR should return to 0x4000 after 32767 steps"
-        );
-    }
-
-    #[test]
-    fn test_ch4_lfsr_7bit_maximal_length() {
-        // Verify the 7-bit Galois LFSR has maximum period of 127 (= 2^7 - 1).
-        let mut lfsr = 0x0040_u16;
-        for _ in 0..127 {
-            let carry = lfsr & 1;
-            lfsr >>= 1;
-            if carry != 0 {
-                lfsr ^= 0x0060;
-            }
-        }
-        assert_eq!(
-            lfsr, 0x0040,
-            "7-bit LFSR should return to 0x0040 after 127 steps"
-        );
     }
 
     // ── sample_ready / get_sample / set_sample_rate ──────────────────────────
